Extract delete modal
